feature/userController (backend/controller): I have organized the controllers to use separately
