+++ server/settings/routes/user.routes.js
@@ -1,14 +1,18 @@
 // Third Party Dependencies.
 const { Router } = require("express");
 // Local Dependencies.
-const User = require("../../../database/models/user.model");
-const { registroUser, login, confirmar, googlelogin, olvidePasswordUser, comprobarToken, nuevoPassword, mostrarUser } = require("../controllers/user/user.controllers");
 const verifyToken = require("../middleware/user/jwt");
+const registroUser = require("../controllers/user/register.controllers");
+const login = require("../controllers/user/login.controllers");
+const googlelogin = require("../controllers/user/googleLogin.controllers");
+const confirmar = require("../controllers/user/confirmAccount.controllers");
+const olvidePasswordUser = require("../controllers/user/forgetPassword.controllers");
+const comprobarToken = require("../controllers/user/findOutToken.controllers");
+const nuevoPassword = require("../controllers/user/newPassword.controllers");
 
 // Router Instance.
 const userRoutes = Router();
 
-userRoutes.get("/", verifyToken, mostrarUser);
 // Registrar Usuarios.
 userRoutes.post("/register", registroUser);
 // Iniciar seccion despues de registrase
